Added different colors for each layer. Right now, they are not able to be changed but in the future, it is easy to add functionality for it.
Convert some types over to enum classes (C++11). Personally, I prefer
these over regular enums because the same name can be defined in
different enums. Also, are strongly typed which is better coding
practice.

Added functionality to current tracing radio buttons and visible tracing
layers checkboxes. These are also stored in the QUndoStack and so can be
undone/redone.

+++ coronalslicewidget.cpp
@@ -75,7 +75,7 @@ void CoronalSliceWidget::setDisplayType(SliceDisplayType type)
     // If the display type is out of the acceptable range, then do nothing
     if (type < SliceDisplayType::FatOnly || type > SliceDisplayType::WaterFraction)
     {
-        qDebug() << "Invalid display type was specified for CoronalSliceWidget: " << type;
+        qDebug() << "Invalid display type was specified for CoronalSliceWidget: " << (int)type;
         return;
     }
